requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.28.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0
Pillow>=10.0.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
